apps/api/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Authentication
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# Job Scraping
python-jobspy>=1.1.0

# Resume Parsing
jsonify-resume>=0.2.0
python-docx>=1.1.0
pypdf>=3.17.0

# OpenAI
openai>=1.10.0

# Scheduling
apscheduler>=3.10.4

# Security
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
